rename the directory of assetSection to assetsSection which corrupted a few imports
