+++ next/pages/pages/[...fullPath].tsx
@@ -20,7 +20,7 @@ import ProcedureTabs from '@/components/molecules/ProcedureTabs'
 import Section from '@/components/molecules/Section'
 import AccordionGroupSection from '@/components/sections/AccordionGroupSection'
 import ArticleListing from '@/components/sections/ArticleListing/ArticleListing'
-import AssetsSection from '@/components/sections/AssetSection/AssetsSection'
+import AssetsSection from '@/components/sections/AssetsSection/AssetsSection'
 import BundleListingSection from '@/components/sections/BundleListingSection'
 import BundleListingSimpleSection from '@/components/sections/BundleListingSimpleSection'
 import CardSection from '@/components/sections/CardSection'
